requirements.txt
bittensor==9.0.0
python-dotenv